Added more Disk I/O operations.
Switched to an ever smaller identifier naming convention. All characters >=6 are shortened, if possible without loss of documentation.
Added buffer utility functions.
Clearing the screen in the boot loader, before jumping to kernal.

+++ kernal/src/io/cli.h
@@ -4,4 +4,4 @@
 
 void cli_write(char *);
 void cli_write_int(int16_t);
-void cli_read(char *, char, uint16_t);
+void cli_read(char *, char, uintptr_t);
